Matrixクラス修正
